Fix icecream in Clear Linux
First, /etc/ld.so.cache doesn't exist (it's in /var/cache). Second the
horrible shell script code that meant to find the base library doesn't
work if we don't use /lib64. But instead of shipping the base library
only, let's ship all the libraries, so other SKX systems in the farm
can run a faster libm.

Signed-off-by: Thiago Macieira <thiago.macieira@intel.com>